PhaseBImplementation/python-server/requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
opencv-python==4.5.5.64
tensorflow==2.17.0
torch
torchvision
numpy
matplotlib
Pillow
OpenEXR
PyQt5
PySide2